# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ElektroPlan",
    "description": "Modulare Plattform fuer Elektrofachbetriebe. Geldbetraege und Mengen werden als Dezimal-String uebertragen (ADR 0005).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/customers": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Kunden auflisten",
        "description": "Kunden der eigenen Organisation - gefiltert, sortiert, nummerierte Seiten.\n\nEine Seite hinter der letzten liefert die letzte vorhandene Seite; das\nFeld ``page`` der Antwort nennt sie (ADR 0017).",
        "operationId": "listCustomers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name, Nummer oder Ort",
              "title": "Q"
            },
            "description": "Name, Nummer oder Ort"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "private",
                    "company"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "created_at",
                "name"
              ],
              "type": "string",
              "default": "created_at",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 1,
              "description": "Seite, beginnend bei 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Seite, beginnend bei 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberedPage_CustomerOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customers"
        ],
        "summary": "Kunden anlegen",
        "description": "Legt einen Kunden an. Die Kundennummer vergibt der Nummernkreis.",
        "operationId": "createCustomer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Kunde",
        "description": "Ein Kunde der eigenen Organisation.",
        "operationId": "getCustomer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "customers"
        ],
        "summary": "Kunde bearbeiten",
        "description": "Aendert einzelne Felder. Die Kundennummer bleibt unveraendert.",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Versionskonflikt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match fehlt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "customers"
        ],
        "summary": "Kunden endgueltig loeschen",
        "description": "Loescht den Kunden **physisch** aus der operativen Datenbank (ADR 0020).\n\nNur Administratoren, und nur wenn dem Kunden **kein einziges** Projekt\nzugeordnet ist - gleich in welchem Status.\nSonst ``409 customer-has-projects``.\n\n**Sperre, Pruefung und Loeschung liegen in einer Transaktion.** Die\nKundenzeile wird zuerst gesperrt; eine gleichzeitige Projektanlage oder ein\nKundenwechsel sperrt dieselbe Zeile und wartet. Danach findet sie den\nKunden nicht mehr (``404``) - ein Projekt ohne Kunden kann nicht entstehen.\n\nDas Protokoll haelt nur Kunden-ID und Kundennummer fest, keine\npersonenbezogenen Daten. Die Kundennummer wird nie wiederverwendet.\nBackups enthalten den Datensatz bis zum Ablauf ihrer Aufbewahrungsfrist\n(docs/security.md, Abschnitt 13).",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Berechtigung fehlt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Versionskonflikt oder dem Kunden sind Projekte zugeordnet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match fehlt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "private",
              "company"
            ],
            "title": "Kind",
            "default": "private"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "billing_street": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Street"
          },
          "billing_postal_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Postal Code"
          },
          "billing_city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing City"
          },
          "billing_country_code": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Billing Country Code",
            "default": "DE"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CustomerCreate",
        "description": "Neuer Kunde. Die Kundennummer vergibt der Nummernkreis."
      },
      "CustomerOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_number": {
            "type": "string",
            "title": "Customer Number"
          },
          "kind": {
            "type": "string",
            "enum": [
              "private",
              "company"
            ],
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "billing_street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Street"
          },
          "billing_postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Postal Code"
          },
          "billing_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing City"
          },
          "billing_country_code": {
            "type": "string",
            "title": "Billing Country Code"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserReference"
          },
          "updated_by": {
            "$ref": "#/components/schemas/UserReference"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_number",
          "kind",
          "name",
          "contact_person",
          "email",
          "phone",
          "billing_street",
          "billing_postal_code",
          "billing_city",
          "billing_country_code",
          "version",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by"
        ],
        "title": "CustomerOut"
      },
      "CustomerUpdate": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "private",
                  "company"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "billing_street": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Street"
          },
          "billing_postal_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Postal Code"
          },
          "billing_city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing City"
          },
          "billing_country_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Country Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CustomerUpdate",
        "description": "Teilaenderung. Nicht gesetzte Felder bleiben unveraendert.\n\nDie Kundennummer ist nicht aenderbar: Sie ist der Bezug bestehender\nBelege (docs/security.md, Abschnitt 13, Punkt 4)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NumberedPage_CustomerOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CustomerOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Tatsaechlich gelieferte Seite, beginnend bei 1. Lag die angefragte Seite hinter der letzten, ist es die letzte vorhandene Seite."
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items",
            "description": "Treffer unter allen Filtern, nur eigener Betrieb."
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "0 bei leerer Treffermenge."
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "NumberedPage[CustomerOut]"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "https://elektroplan.internal/errors/not-found"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "organizations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrganizationChoice"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizations"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "Antwortkoerper nach RFC 9457."
      },
...
      "UserReference": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "member",
              "unknown",
              "system"
            ],
            "title": "Kind"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "UserReference",
        "description": "Bearbeiter eines Datensatzes, wie die Oberflaeche ihn zeigen darf."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}